- alteração da tabela configuracoes_padrao referente coluna "rel_maxlinhas" - alteração do programa de "configuracao_padrao" referente alteração da tabela - página principal deixa de mostrar informações a "anonymous users"
git-svn-id: http://svn.softwarepublico.gov.br/svn/cacic/cacic/trunk/gerente@544 fecfc0c7-e812-0410-ae72-849f08638ee7

+++ instalador/sql/cacic_v2.2.2.sql
@@ -63,7 +63,7 @@ ALTER TABLE configuracoes_locais
 
 ALTER TABLE configuracoes_padrao
     ADD te_exibe_graficos varchar(100) NOT NULL DEFAULT '[acessos_locais][so][acessos][locais]' COMMENT '' AFTER id_default_body_bgcolor,
-    ADD nu_rel_maxlinhas int(3) default 50 AFTER nu_exec_apos,
+    ADD nu_rel_maxlinhas smallint unsigned default 50 AFTER nu_exec_apos,
     ENGINE=InnoDB CHARACTER SET=latin1;
